bug finishing

+++ controllers/keyController.js
@@ -1,7 +1,6 @@
 import { keyModel } from "../models/keyModels.js";
 import { languageModels } from "../models/languageModels.js";
 import { pageModels } from "../models/pageModels.js";
-import { websiteModels } from "../models/websiteModels.js";
 import logger from "../utilities/logger.js";
 import message from "../utilities/messages/message.js";
 import { sendBadRequest, sendSuccess } from "../utilities/response/index.js";
@@ -11,7 +10,7 @@ import { sendBadRequest, sendSuccess } from "../utilities/response/index.js";
 export const createLanguageKey = async (req, res) => {
     try {
         const data = req.body
-        const pageData = await pageModels.findOne({ _id: req.params.pageid })
+        const pageData = await pageModels.findOne({ _id: req.params.pageId })
         if (!pageData) return sendBadRequest(res, message.pageDataNotFound)
         if (pageData.status === false) {
             return sendBadRequest(res, message.pageIsNotLongerExist)
@@ -79,7 +78,7 @@ export const getAllKeyData = async (req, res) => {
         if (req.query.status) {
             options.status = req.query.status
         }
-        const languageData = await keyModel.find(options).populate("page_id", "name")
+        const languageData = await keyModel.find(options).populate("page_id", "name").populate('language.lg', 'name')
         if (!languageData) {
             return sendBadRequest(res, message.languageDataNotFound)
         }
@@ -97,7 +96,7 @@ export const getAllKeyData = async (req, res) => {
 export const getAllKeyDataByPageId = async (req, res) => {
     try {
         const options = {}
-        options.page_id = req.params.pageid
+        options.page_id = req.params.pageId
         if (req.query.status) {
             options.status = req.query.status
         }
@@ -120,7 +119,7 @@ export const getKeyDataByKeyId = async (req, res) => {
         if (req.query.status) {
             options.status = req.query.status
         }
-        const keyData = await keyModel.find({ _id: req.params.keyid }).select({ name: 1 })
+        const keyData = await keyModel.find({ _id: req.params.keyId }).select({ name: 1 })
         if (!keyData) {
             return sendBadRequest(res, message.keyDataNotFound)
         }
@@ -141,7 +140,7 @@ export const getKeyDataByKeyId = async (req, res) => {
 export const updateKeyData = async (req, res) => {
     try {
         const data = req.body
-        let keyData = await keyModel.findOne({ _id: req.params.keyid })
+        let keyData = await keyModel.findOne({ _id: req.params.keyId })
         if (!keyData) {
             return sendBadRequest(res, message.keyDataNotFound)
         }
@@ -200,13 +199,18 @@ export const updateKeyData = async (req, res) => {
 //use for delete key data
 export const deleteKeyData = async (req, res) => {
     try {
-        const keyData = await keyModel.findOne({ _id: req.params.keyid })
+        const keyData = await keyModel.findOne({ _id: req.params.keyId })
         if (!keyData) {
             return sendBadRequest(res, message.keyDataNotFound)
         }
 
         if (!req.website.pages.includes(keyData.page_id)) {
             return sendBadRequest(res, message.pageDataNotFound)
+        }
+
+        const pageData = await pageModels.findOne({ keys: { $in: keyData._id } })
+        if (!pageData) {
+          return sendBadRequest(res, message.pageDataNotFound)
         }
         await pageData.keys.pull(keyData._id)
         await keyData.delete()
